fix: rename examples
