+++ examples/index.html
@@ -3,13 +3,12 @@
 <head>
 <meta charset="utf-8"/>
 <meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no">
-<title>react-web-pull-to-refresh examples</title>
+<title>PullDownToRefresh examples</title>
 </head>
 <body>
-  <h1>react-web-pull-to-refresh examples</h1>
+  <h1>PullDownToRefresh examples</h1>
   <ul>
-    <li><a href="/string/">String example</a></li>
-    <li><a href="/icon/">Icon example</a></li>
+    <li><a href="/pull_down_to_refresh">PullDownToRefresh example</a></li>
   </ul>
 </body>
 </html>
