+++ examples/pull_down_to_refresh/icon/index.js
