+++ examples/pull_down_to_refresh/index.html
@@ -1,0 +1,15 @@
+<!doctype html>
+<html>
+<head>
+<meta charset="utf-8"/>
+<meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no">
+<title>pull-down-to-refresh examples</title>
+</head>
+<body>
+  <h1>pull-down-to-refresh examples</h1>
+  <ul>
+    <li><a href="/pull_down_to_refresh/string/">String example</a></li>
+    <li><a href="/pull_down_to_refresh/icon/">Icon example</a></li>
+  </ul>
+</body>
+</html>
