+++ examples/pull_down_to_refresh/string/index.js
